Move layouts in /src Fixes #289
