+++ lib/src/layouts/yaru_detail_page.dart
